+++ api/QCVOC.Api/Services/Data/Repository/ServiceRepository.cs
@@ -3,7 +3,7 @@
 //     in the project root for full license information.
 // </copyright>
 
-namespace QCVOC.Api.Service.Data.Repository
+namespace QCVOC.Api.Services.Data.Repository
 {
     using System;
     using System.Collections.Generic;
@@ -12,7 +12,7 @@ namespace QCVOC.Api.Service.Data.Repository
     using QCVOC.Api.Common;
     using QCVOC.Api.Common.Data.ConnectionFactory;
     using QCVOC.Api.Common.Data.Repository;
-    using QCVOC.Api.Service.Data.Model;
+    using QCVOC.Api.Services.Data.Model;
 
     /// <summary>
     ///     Provides data access for <see cref="Service"/>.
